backend/requirements.txt
django-cors-headers==3.7.0
django-filter==2.4.0
django-multiselectfield==0.1.12
djoser==2.1.0
drf-yasg==1.20.0
gunicorn==20.0.4
pillow==8.1.2
pip-chill==1.0.1
